+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Yourloops data visualization for diabetes device data
 Library for Tidepool's timeline-style diabetes data visualization(s) used in Blip and was forked from tidepool/tideline.
 
+## Unreleased
+### Fixed 
+- PT-1205 Add timezone info on tooltips when necessary
+
 ## 1.9.2 - 2020-04-17
 ### Fixed
 - PT-1231 Basics: Total basal events count does not give the correct total